refactor(bayn): extract qualification decision kernel
